requirements.txt
flask
mysql-connector-python
python-dotenv
Flask-Mail